requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
httpx>=0.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
aiocache>=0.12.0
prometheus-fastapi-instrumentator>=7.0.0
fastapi-limiter>=0.1.5
redis>=5.0.0,<6.0.0
python-multipart>=0.0.5
psycopg2-binary>=2.9.0
ruff>=0.3.0
